requirements.txt
# Core scientific computing
numpy>=1.24
pandas>=2.0
scipy>=1.10

# Chemistry / molecular
rdkit>=2023.3.1
biopython>=1.81

# Machine learning
torch>=2.0
torch-geometric>=2.3

# Docking
vina>=1.2.3

# Visualization
matplotlib>=3.7
seaborn>=0.12
py3Dmol>=2.0
ipywidgets>=8.0

# Dimensionality reduction
umap-learn>=0.5
scikit-learn>=1.3

# Notebook / utilities
jupyter>=1.0
jupyterlab>=4.0
tqdm>=4.65
pyyaml>=6.0
requests>=2.31
